+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fix bug in TripListConverter when estimating trip durations for maximum duration activities
 - Improve/add tests for ScheduleWaitingTimeEstimator
 - Add Apollo reader
+- fix int overflow exceptions for tours with many trips in DefaultModeChainGenerator
 
 **1.0.9**